+++ crates/services/executor/src/executor.rs
@@ -134,6 +134,8 @@ use fuel_core_types::{
             TransactionExecutionStatus,
             TransactionValidityError,
             UncommittedResult,
+            UncommittedValidationResult,
+            ValidationResult,
         },
         relayer::Event,
     },
@@ -247,24 +249,63 @@ where
         let ExecutionData {
             message_ids,
             event_inbox_root,
+            changes,
+            events,
+            tx_status,
+            skipped_transactions,
+            coinbase,
+            used_gas,
             ..
-        } = &execution_data;
+        } = execution_data;
 
         let block = partial_block
-            .generate(&message_ids[..], *event_inbox_root)
+            .generate(&message_ids[..], event_inbox_root)
             .map_err(ExecutorError::BlockHeaderError)?;
-        Self::uncommitted_block_results(block, execution_data)
+
+        let finalized_block_id = block.id();
+
+        debug!(
+            "Block {:#x} fees: {} gas: {}",
+            finalized_block_id, coinbase, used_gas
+        );
+
+        let result = ExecutionResult {
+            block,
+            skipped_transactions,
+            tx_status,
+            events,
+        };
+
+        Ok(UncommittedResult::new(result, changes))
     }
 
     pub fn validate_without_commit(
         self,
-        block: Block,
-    ) -> ExecutorResult<UncommittedResult<Changes>> {
+        block: &Block,
+    ) -> ExecutorResult<UncommittedValidationResult<Changes>> {
         let consensus_params_version = block.header().consensus_parameters_version;
         let (block_executor, storage_tx) =
             self.into_executor(consensus_params_version)?;
-        let execution_data = block_executor.validate_block(&block, storage_tx)?;
-        Self::uncommitted_block_results(block, execution_data)
+
+        let ExecutionData {
+            coinbase,
+            used_gas,
+            tx_status,
+            events,
+            changes,
+            ..
+        } = block_executor.validate_block(block, storage_tx)?;
+
+        let finalized_block_id = block.id();
+
+        debug!(
+            "Block {:#x} fees: {} gas: {}",
+            finalized_block_id, coinbase, used_gas
+        );
+
+        let result = ValidationResult { tx_status, events };
+
+        Ok(UncommittedValidationResult::new(result, changes))
     }
 
     fn into_executor(
@@ -284,37 +325,6 @@ where
             .into_owned();
         let executor = BlockExecutor::new(self.relayer, self.options, consensus_params)?;
         Ok((executor, storage_tx))
-    }
-
-    fn uncommitted_block_results(
-        block: Block,
-        data: ExecutionData,
-    ) -> ExecutorResult<UncommittedResult<Changes>> {
-        let ExecutionData {
-            coinbase,
-            used_gas,
-            tx_status,
-            skipped_transactions,
-            events,
-            changes,
-            ..
-        } = data;
-
-        let finalized_block_id = block.id();
-
-        debug!(
-            "Block {:#x} fees: {} gas: {}",
-            finalized_block_id, coinbase, used_gas
-        );
-
-        let result = ExecutionResult {
-            block,
-            skipped_transactions,
-            tx_status,
-            events,
-        };
-
-        Ok(UncommittedResult::new(result, changes))
     }
 }