+++ crates/services/importer/src/importer/test.rs
@@ -22,7 +22,6 @@ use fuel_core_types::{
         consensus::Consensus,
         SealedBlock,
     },
-    fuel_tx::TxId,
     fuel_types::BlockHeight,
     services::{
         block_importer::{
@@ -31,8 +30,9 @@ use fuel_core_types::{
         },
         executor::{
             Error as ExecutorError,
-            ExecutionResult,
             Result as ExecutorResult,
+            UncommittedValidationResult,
+            ValidationResult,
         },
         Uncommitted,
     },
@@ -65,12 +65,6 @@ fn u32_to_merkle_root(number: u32) -> MerkleRoot {
     let mut root = [0; 32];
     root[0..4].copy_from_slice(&number.to_be_bytes());
     MerkleRoot::from(root)
-}
-
-#[derive(Clone, Debug)]
-struct MockExecutionResult {
-    block: SealedBlock,
-    skipped_transactions: usize,
 }
 
 fn genesis(height: u32) -> SealedBlock {
@@ -145,11 +139,14 @@ fn storage_failure_error() -> Error {
     storage_failure::<()>().unwrap_err().into()
 }
 
-fn ex_result(height: u32, skipped_transactions: usize) -> MockExecutionResult {
-    MockExecutionResult {
-        block: poa_block(height),
-        skipped_transactions,
-    }
+fn ex_result() -> ExecutorResult<UncommittedValidationResult<Changes>> {
+    Ok(Uncommitted::new(
+        ValidationResult {
+            tx_status: vec![],
+            events: vec![],
+        },
+        Default::default(),
+    ))
 }
 
 fn execution_failure<T>() -> ExecutorResult<T> {
@@ -162,24 +159,10 @@ fn execution_failure_error() -> Error {
 
 fn executor<R>(result: R) -> MockValidator
 where
-    R: Fn() -> ExecutorResult<MockExecutionResult> + Send + 'static,
+    R: Fn() -> ExecutorResult<UncommittedValidationResult<Changes>> + Send + 'static,
 {
     let mut executor = MockValidator::default();
-    executor.expect_validate().return_once(move |_| {
-        let mock_result = result()?;
-        let skipped_transactions: Vec<_> = (0..mock_result.skipped_transactions)
-            .map(|_| (TxId::zeroed(), ExecutorError::BlockMismatch))
-            .collect();
-        Ok(Uncommitted::new(
-            ExecutionResult {
-                block: mock_result.block.entity,
-                skipped_transactions,
-                tx_status: vec![],
-                events: vec![],
-            },
-            Default::default(),
-        ))
-    });
+    executor.expect_validate().return_once(move |_| result());
 
     executor
 }
@@ -329,7 +312,6 @@ async fn commit_result_and_execute_and_commit_poa(
 ) -> Result<(), Error> {
     // `execute_and_commit` and `commit_result` should have the same
     // validation rules(-> test cases) during committing the result.
-    let height = *sealed_block.entity.header().height();
     let transaction = db_transaction();
     let commit_result =
         commit_result_assert(sealed_block.clone(), underlying_db(), transaction).await;
@@ -339,7 +321,7 @@ async fn commit_result_and_execute_and_commit_poa(
     let execute_and_commit_result = execute_and_commit_assert(
         sealed_block,
         db,
-        executor(ok(ex_result(height.into(), 0))),
+        executor(ex_result),
         verifier(ok(())),
     )
     .await;
@@ -449,24 +431,19 @@ fn one_lock_at_the_same_time() {
 
 ///////// New block, Block After Execution, Verification result, commits /////////
 #[test_case(
-    genesis(113), ok(ex_result(113, 0)), ok(()), 0
+    genesis(113), ex_result, ok(()), 0
     => Err(Error::ExecuteGenesis);
     "cannot execute genesis block"
 )]
 #[test_case(
-    poa_block(1), ok(ex_result(1, 0)), ok(()), 1
+    poa_block(1), ex_result, ok(()), 1
     => Ok(());
     "commits block 1"
 )]
 #[test_case(
-    poa_block(113), ok(ex_result(113, 0)), ok(()), 1
+    poa_block(113), ex_result, ok(()), 1
     => Ok(());
     "commits block 113"
-)]
-#[test_case(
-    poa_block(113), ok(ex_result(114, 0)), ok(()), 0
-    => Err(Error::BlockIdMismatch(poa_block(113).entity.id(), poa_block(114).entity.id()));
-    "execution result should match block height"
 )]
 #[test_case(
     poa_block(113), execution_failure, ok(()), 0
@@ -474,12 +451,7 @@ fn one_lock_at_the_same_time() {
     "commit fails if execution fails"
 )]
 #[test_case(
-    poa_block(113), ok(ex_result(113, 1)), ok(()), 0
-    => Err(Error::SkippedTransactionsNotEmpty);
-    "commit fails if there are skipped transactions"
-)]
-#[test_case(
-    poa_block(113), ok(ex_result(113, 0)), verification_failure, 0
+    poa_block(113), ex_result, verification_failure, 0
     => Err(verification_failure_error());
     "commit fails if verification fails"
 )]
@@ -491,7 +463,10 @@ async fn execute_and_commit_and_verify_and_execute_block_poa<V, P>(
     commits: usize,
 ) -> Result<(), Error>
 where
-    P: Fn() -> ExecutorResult<MockExecutionResult> + Send + Clone + 'static,
+    P: Fn() -> ExecutorResult<UncommittedValidationResult<Changes>>
+        + Send
+        + Clone
+        + 'static,
     V: Fn() -> anyhow::Result<()> + Send + Clone + 'static,
 {
     // `execute_and_commit` and `verify_and_execute_block` should have the same
@@ -527,7 +502,7 @@ fn verify_and_execute_assert<P, V>(
     verifier_result: V,
 ) -> Result<(), Error>
 where
-    P: Fn() -> ExecutorResult<MockExecutionResult> + Send + 'static,
+    P: Fn() -> ExecutorResult<UncommittedValidationResult<Changes>> + Send + 'static,
     V: Fn() -> anyhow::Result<()> + Send + 'static,
 {
     let importer = Importer::default_config(
@@ -543,7 +518,7 @@ where
 fn verify_and_execute_allowed_when_locked() {
     let importer = Importer::default_config(
         MockDatabase::default(),
-        executor(ok(ex_result(13, 0))),
+        executor(ex_result),
         verifier(ok(())),
     );