+++ crates/services/upgradable-executor/src/instance.rs
@@ -31,9 +31,9 @@ use fuel_core_types::{
 use fuel_core_wasm_executor::utils::{
     pack_exists_size_result,
     unpack_ptr_and_len,
-    InputType,
+    InputSerializationType,
     ReturnType,
-    WasmExecutionBlockTypes,
+    WasmSerializationBlockTypes,
 };
 use std::{
     collections::HashMap,
@@ -459,8 +459,8 @@ impl Instance<Relayer> {
         components: Components<()>,
         options: ExecutionOptions,
     ) -> ExecutorResult<Instance<InputData>> {
-        let input = InputType::V1 {
-            block: WasmExecutionBlockTypes::Production(components),
+        let input = InputSerializationType::V1 {
+            block: WasmSerializationBlockTypes::Production(components),
             options,
         };
         self.add_input_data(input)
@@ -471,26 +471,30 @@ impl Instance<Relayer> {
         components: Components<()>,
         options: ExecutionOptions,
     ) -> ExecutorResult<Instance<InputData>> {
-        let input = InputType::V1 {
-            block: WasmExecutionBlockTypes::DryRun(components),
+        let input = InputSerializationType::V1 {
+            block: WasmSerializationBlockTypes::DryRun(components),
             options,
         };
         self.add_input_data(input)
     }
 
+    //
     pub fn add_validation_input_data(
         self,
-        block: Block,
+        block: &Block,
         options: ExecutionOptions,
     ) -> ExecutorResult<Instance<InputData>> {
-        let input = InputType::V1 {
-            block: WasmExecutionBlockTypes::Validation(block),
+        let input = InputSerializationType::V1 {
+            block: WasmSerializationBlockTypes::Validation(block),
             options,
         };
         self.add_input_data(input)
     }
 
-    fn add_input_data(mut self, input: InputType) -> ExecutorResult<Instance<InputData>> {
+    fn add_input_data(
+        mut self,
+        input: InputSerializationType,
+    ) -> ExecutorResult<Instance<InputData>> {
         let encoded_input = postcard::to_allocvec(&input).map_err(|e| {
             ExecutorError::Other(format!(
                 "Failed encoding of the input for `input` function: {}",